add long int

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -8,7 +8,7 @@
 size_t print_listint_safe(const listint_t *head)
 {
 	size_t i = 0;
-	int len;
+	long int len;
 
 	if (head == NULL)
 	{
